requirements.txt
llama-index
chromadb
PyMuPDF
python-dotenv
tqdm
sentence-transformers
transformers
accelerate
jupyterlab
